refactor (gravsearch): move method createMainQuery for fulltext search to GravsearchUtilsV2
