+++ webapi/src/main/scala/org/knora/webapi/responders/v2/SearchResponderV2.scala
@@ -30,10 +30,9 @@ import org.knora.webapi.util.IriConversions._
 import org.knora.webapi.util._
 import org.knora.webapi.util.search.ApacheLuceneSupport.{CombineSearchTerms, MatchStringWhileTyping}
 import org.knora.webapi.util.search._
-import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.Gravsearch.GravsearchConstants
-import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.Gravsearch._
 import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.SparqlTransformation._
 import org.knora.webapi.util.search.gravsearch._
+
 import scala.concurrent.Future
 
 /**
@@ -129,97 +128,10 @@ class SearchResponderV2 extends ResponderWithStandoffV2 {
       */
     private def fulltextSearchV2(searchValue: String, offset: Int, limitToProject: Option[IRI], limitToResourceClass: Option[SmartIri], limitToStandoffClass: Option[SmartIri], requestingUser: UserADM): Future[ReadResourcesSequenceV2] = {
 
+        import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.FulltextSearch._
         import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.FulltextSearch.FullTextSearchConstants._
 
         val groupConcatSeparator = StringFormatter.INFORMATION_SEPARATOR_ONE
-
-        /**
-          * Creates a CONSTRUCT query for the given resource and value object IRIs.
-          *
-          * @param resourceIris    the IRIs of the resources to be queried.
-          * @param valueObjectIris the IRIs of the value objects to be queried.
-          * @return a [[ConstructQuery]].
-          */
-        def createMainQuery(resourceIris: Set[IRI], valueObjectIris: Set[IRI]): ConstructQuery = {
-
-            // WHERE patterns for the resources: check that the resource are a knora-base:Resource and that it is not marked as deleted
-            val wherePatternsForResources = Seq(
-                ValuesPattern(resourceVar, resourceIris.map(iri => IriRef(iri.toSmartIri))), // a ValuePattern that binds the resource IRIs to the resource variable
-                StatementPattern.makeInferred(subj = resourceVar, pred = IriRef(OntologyConstants.Rdf.Type.toSmartIri), obj = IriRef(OntologyConstants.KnoraBase.Resource.toSmartIri)),
-                StatementPattern.makeExplicit(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.IsDeleted.toSmartIri), obj = XsdLiteral(value = "false", datatype = OntologyConstants.Xsd.Boolean.toSmartIri)),
-                StatementPattern.makeExplicit(subj = resourceVar, pred = resourcePropVar, obj = resourceObjectVar)
-            )
-
-            //  mark resources as the main resource and a knora-base:Resource in CONSTRUCT clause and return direct assertions about resources
-            val constructPatternsForResources = Seq(
-                StatementPattern(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.IsMainResource.toSmartIri), obj = XsdLiteral(value = "true", datatype = OntologyConstants.Xsd.Boolean.toSmartIri)),
-                StatementPattern(subj = resourceVar, pred = IriRef(OntologyConstants.Rdf.Type.toSmartIri), obj = IriRef(OntologyConstants.KnoraBase.Resource.toSmartIri)),
-                StatementPattern(subj = resourceVar, pred = resourcePropVar, obj = resourceObjectVar)
-            )
-
-            if (valueObjectIris.nonEmpty) {
-                // value objects are to be queried
-
-                // WHERE patterns for statements about the resources' values
-                val wherePatternsForValueObjects = Seq(
-                    ValuesPattern(resourceValueObject, valueObjectIris.map(iri => IriRef(iri.toSmartIri))),
-                    StatementPattern.makeInferred(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.HasValue.toSmartIri), obj = resourceValueObject),
-                    StatementPattern.makeExplicit(subj = resourceVar, pred = resourceValueProp, obj = resourceValueObject),
-                    StatementPattern.makeExplicit(subj = resourceValueObject, pred = IriRef(OntologyConstants.KnoraBase.IsDeleted.toSmartIri), obj = XsdLiteral(value = "false", datatype = OntologyConstants.Xsd.Boolean.toSmartIri)),
-                    StatementPattern.makeExplicit(subj = resourceValueObject, pred = resourceValueObjectProp, obj = resourceValueObjectObj)
-                )
-
-                // return assertions about value objects
-                val constructPatternsForValueObjects = Seq(
-                    StatementPattern(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.HasValue.toSmartIri), obj = resourceValueObject),
-                    StatementPattern(subj = resourceVar, pred = resourceValueProp, obj = resourceValueObject),
-                    StatementPattern(subj = resourceValueObject, pred = resourceValueObjectProp, obj = resourceValueObjectObj)
-                )
-
-                // WHERE patterns for standoff belonging to value objects (if any)
-                val wherePatternsForStandoff = Seq(
-                    ValuesPattern(resourceValueObject, valueObjectIris.map(iri => IriRef(iri.toSmartIri))),
-                    StatementPattern.makeExplicit(subj = resourceValueObject, pred = IriRef(OntologyConstants.KnoraBase.ValueHasStandoff.toSmartIri), obj = standoffNodeVar),
-                    StatementPattern.makeExplicit(subj = standoffNodeVar, pred = standoffPropVar, obj = standoffValueVar)
-                )
-
-                // return standoff
-                val constructPatternsForStandoff = Seq(
-                    StatementPattern(subj = resourceValueObject, pred = IriRef(OntologyConstants.KnoraBase.ValueHasStandoff.toSmartIri), obj = standoffNodeVar),
-                    StatementPattern(subj = standoffNodeVar, pred = standoffPropVar, obj = standoffValueVar)
-                )
-
-                ConstructQuery(
-                    constructClause = ConstructClause(
-                        statements = constructPatternsForResources ++ constructPatternsForValueObjects ++ constructPatternsForStandoff
-                    ),
-                    whereClause = WhereClause(
-                        Seq(
-                            UnionPattern(
-                                Seq(wherePatternsForResources, wherePatternsForValueObjects, wherePatternsForStandoff)
-                            )
-                        )
-                    )
-                )
-
-            } else {
-                // no value objects are to be queried
-
-                ConstructQuery(
-                    constructClause = ConstructClause(
-                        statements = constructPatternsForResources
-                    ),
-                    whereClause = WhereClause(
-                        Seq(
-                            UnionPattern(
-                                Seq(wherePatternsForResources)
-                            )
-                        )
-                    )
-                )
-            }
-
-        }
 
         val searchTerms: CombineSearchTerms = CombineSearchTerms(searchValue)
 
@@ -447,6 +359,8 @@ class SearchResponderV2 extends ResponderWithStandoffV2 {
       * @return a [[ReadResourcesSequenceV2]] representing the resources that have been found.
       */
     private def gravsearchV2(inputQuery: ConstructQuery, requestingUser: UserADM): Future[ReadResourcesSequenceV2] = {
+
+        import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.Gravsearch._
 
         for {
             // Do type inspection and remove type annotations from the WHERE clause.
