+++ webapi/src/main/scala/org/knora/webapi/util/search/gravsearch/GravsearchUtilV2.scala
@@ -25,7 +25,6 @@ import org.knora.webapi.messages.store.triplestoremessages.{SparqlConstructRespo
 import org.knora.webapi.util.{ConstructResponseUtilV2, ErrorHandlingMap, SmartIri, StringFormatter}
 import org.knora.webapi.util.search._
 import org.knora.webapi.util.IriConversions._
-import org.knora.webapi.util.search.gravsearch.GravsearchUtilV2.SparqlTransformation.{ResourceIrisAndValueObjectIris, createUniqueVariableNameFromEntityAndProperty, traverseValuePropertyAssertions}
 
 
 /**
@@ -395,6 +394,97 @@ object GravsearchUtilV2 {
             // SPARQL variable representing the objects of a standoff node of a (text) value object
             val standoffValueVar: QueryVariable = QueryVariable("standoffValue")
         }
+
+        /**
+          * Creates a CONSTRUCT query for the given resource and value object IRIs.
+          *
+          * @param resourceIris    the IRIs of the resources to be queried.
+          * @param valueObjectIris the IRIs of the value objects to be queried.
+          * @return a [[ConstructQuery]].
+          */
+        def createMainQuery(resourceIris: Set[IRI], valueObjectIris: Set[IRI]): ConstructQuery = {
+
+            import FullTextSearchConstants._
+
+            // WHERE patterns for the resources: check that the resource are a knora-base:Resource and that it is not marked as deleted
+            val wherePatternsForResources = Seq(
+                ValuesPattern(resourceVar, resourceIris.map(iri => IriRef(iri.toSmartIri))), // a ValuePattern that binds the resource IRIs to the resource variable
+                StatementPattern.makeInferred(subj = resourceVar, pred = IriRef(OntologyConstants.Rdf.Type.toSmartIri), obj = IriRef(OntologyConstants.KnoraBase.Resource.toSmartIri)),
+                StatementPattern.makeExplicit(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.IsDeleted.toSmartIri), obj = XsdLiteral(value = "false", datatype = OntologyConstants.Xsd.Boolean.toSmartIri)),
+                StatementPattern.makeExplicit(subj = resourceVar, pred = resourcePropVar, obj = resourceObjectVar)
+            )
+
+            //  mark resources as the main resource and a knora-base:Resource in CONSTRUCT clause and return direct assertions about resources
+            val constructPatternsForResources = Seq(
+                StatementPattern(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.IsMainResource.toSmartIri), obj = XsdLiteral(value = "true", datatype = OntologyConstants.Xsd.Boolean.toSmartIri)),
+                StatementPattern(subj = resourceVar, pred = IriRef(OntologyConstants.Rdf.Type.toSmartIri), obj = IriRef(OntologyConstants.KnoraBase.Resource.toSmartIri)),
+                StatementPattern(subj = resourceVar, pred = resourcePropVar, obj = resourceObjectVar)
+            )
+
+            if (valueObjectIris.nonEmpty) {
+                // value objects are to be queried
+
+                // WHERE patterns for statements about the resources' values
+                val wherePatternsForValueObjects = Seq(
+                    ValuesPattern(resourceValueObject, valueObjectIris.map(iri => IriRef(iri.toSmartIri))),
+                    StatementPattern.makeInferred(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.HasValue.toSmartIri), obj = resourceValueObject),
+                    StatementPattern.makeExplicit(subj = resourceVar, pred = resourceValueProp, obj = resourceValueObject),
+                    StatementPattern.makeExplicit(subj = resourceValueObject, pred = IriRef(OntologyConstants.KnoraBase.IsDeleted.toSmartIri), obj = XsdLiteral(value = "false", datatype = OntologyConstants.Xsd.Boolean.toSmartIri)),
+                    StatementPattern.makeExplicit(subj = resourceValueObject, pred = resourceValueObjectProp, obj = resourceValueObjectObj)
+                )
+
+                // return assertions about value objects
+                val constructPatternsForValueObjects = Seq(
+                    StatementPattern(subj = resourceVar, pred = IriRef(OntologyConstants.KnoraBase.HasValue.toSmartIri), obj = resourceValueObject),
+                    StatementPattern(subj = resourceVar, pred = resourceValueProp, obj = resourceValueObject),
+                    StatementPattern(subj = resourceValueObject, pred = resourceValueObjectProp, obj = resourceValueObjectObj)
+                )
+
+                // WHERE patterns for standoff belonging to value objects (if any)
+                val wherePatternsForStandoff = Seq(
+                    ValuesPattern(resourceValueObject, valueObjectIris.map(iri => IriRef(iri.toSmartIri))),
+                    StatementPattern.makeExplicit(subj = resourceValueObject, pred = IriRef(OntologyConstants.KnoraBase.ValueHasStandoff.toSmartIri), obj = standoffNodeVar),
+                    StatementPattern.makeExplicit(subj = standoffNodeVar, pred = standoffPropVar, obj = standoffValueVar)
+                )
+
+                // return standoff
+                val constructPatternsForStandoff = Seq(
+                    StatementPattern(subj = resourceValueObject, pred = IriRef(OntologyConstants.KnoraBase.ValueHasStandoff.toSmartIri), obj = standoffNodeVar),
+                    StatementPattern(subj = standoffNodeVar, pred = standoffPropVar, obj = standoffValueVar)
+                )
+
+                ConstructQuery(
+                    constructClause = ConstructClause(
+                        statements = constructPatternsForResources ++ constructPatternsForValueObjects ++ constructPatternsForStandoff
+                    ),
+                    whereClause = WhereClause(
+                        Seq(
+                            UnionPattern(
+                                Seq(wherePatternsForResources, wherePatternsForValueObjects, wherePatternsForStandoff)
+                            )
+                        )
+                    )
+                )
+
+            } else {
+                // no value objects are to be queried
+
+                ConstructQuery(
+                    constructClause = ConstructClause(
+                        statements = constructPatternsForResources
+                    ),
+                    whereClause = WhereClause(
+                        Seq(
+                            UnionPattern(
+                                Seq(wherePatternsForResources)
+                            )
+                        )
+                    )
+                )
+            }
+
+        }
+
     }
 
     object Gravsearch {
@@ -460,6 +550,8 @@ object GravsearchUtilV2 {
           * @return a Set of [[PropertyTypeInfo]] representing the value and link value properties to be returned to the client.
           */
         def collectValueVariablesForResource(constructClause: ConstructClause, resource: Entity, typeInspectionResult: GravsearchTypeInspectionResult, variableConcatSuffix: String): Set[QueryVariable] = {
+
+            import SparqlTransformation._
 
             // make sure resource is a query variable or an IRI
             resource match {
@@ -766,6 +858,8 @@ object GravsearchUtilV2 {
                                                     dependentResourceIrisPerMainResource: DependentResourcesPerMainResource,
                                                     valueObjectVarsAndIrisPerMainResource: ValueObjectVariablesAndValueObjectIris,
                                                     requestingUser: UserADM): Map[IRI, ConstructResponseUtilV2.ResourceWithValueRdfData] = {
+
+            import SparqlTransformation._
 
             // separate main resources and value objects (dependent resources are nested)
             // this method removes resources and values the requesting users has insufficient permissions on (missing view permissions).
